# doc/swagger/v1/swagger.yml
openapi: 3.0.3
info:
  title: Article Management System
  description: >-
    Collection of tests for the article management system API.


    [https://github.com/nathanbizkit/article-management](https://github.com/nathanbizkit/article-management)
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:8000/api/v1
  - url: https://localhost:8443/api/v1
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
    refreshAuth:
      type: apiKey
      in: cookie
      name: refreshToken
security:
  - sessionAuth: []
paths:
...
  /refresh_token:
    post:
      tags:
        - Auth
      summary: Refresh Token
      description: >-
        Refreshes session with authentication in request headers and returns a
        new authentication in cookie.
      operationId: refreshToken
      security:
        - refreshAuth: []
      responses:
        "204":
          description: >
            Successfully refreshed authentication. The authentication is
            returned in cookie (session in cookie named `session` and refresh
            token in cookie named `refreshToken`). You need to include these two
            cookies in subsequent private requests.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: >-
                  session=abcde12345; Path=/; HttpOnly; refreshToken=abcde12345;
                  Path=/; HttpOnly
...
  /profiles/{username}:
    get:
      tags:
        - Profiles
      summary: Profile by Username
      description: Retrieves other user's profile.
      operationId: profile
      responses:
        "200":
          description: A profile object
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  name:
                    type: string
                  bio:
                    type: string
                  image:
                    type: string
                    format: uri
                  following:
                    type: boolean
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
  /profiles/{username}/follow:
    post:
      tags:
        - Profiles
      summary: Follow User
      description: Follows other user.
      operationId: followUser
      responses:
        "200":
          description: A profile object
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  name:
                    type: string
                  bio:
                    type: string
                  image:
                    type: string
                    format: uri
                  following:
                    type: boolean
    delete:
      tags:
        - Profiles
      summary: Unfollow User
      description: Unfollow other user.
      operationId: unfollowUser
      responses:
        "200":
          description: A profile object
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  name:
                    type: string
                  bio:
                    type: string
                  image:
                    type: string
                    format: uri
                  following:
                    type: boolean
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
...